Support multiple highlights in vim. Match non-ASCII.
Prior to this, I had multiple highlights and matches but it never quite
worked. One would match, but not the other. And it changed dependening
on the order in the file.

Turns out, that's how VIM is supposed to work :(
But there is a way! And that's with `matchadd()`:
  * http://vim.wikia.com/wiki/Highlight_multiple_words

Also, I added highlighting of non-ASCII characters as those have a
tendency to blow up Puppet and other software when parsed.